[권미경] #7 - LocalDateTime 통해서 종료일자-시작일자 날짜차이 카운트하기

+++ src/main/java/com/kkomiding/davena/holiday/HolidayController.java
@@ -40,7 +40,7 @@ public class HolidayController {
 							,Model model) {
 		
 		int userId = (Integer)session.getAttribute("userId");
-		int count = holidayService.selectHolidayCount(userId);
+		long count = holidayService.selectHolidayCount(userId);
 		
 		model.addAttribute("count", count);
